move stra_iteri to handle_input and remove parse_utils
